Text_to_Speech_Converter/requirements.txt
nltk==3.8.1
newspaper3k==0.2.8
gtts==2.5.1
